refactor mobilenet_v2 utils
