--- models/demos/mobilenetv2/tests/perf/mobilenetv2_common.py
@@ -1,6 +1,0 @@
-# SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
-
-# SPDX-License-Identifier: Apache-2.0
-
-MOBILENETV2_L1_SMALL_SIZE = 8 * 1024  # 8 KiB
-MOBILENETV2_BATCH_SIZE = 10
