+++ models/demos/mobilenetv2/tests/perf/test_perf_mobilenetv2.py
@@ -6,7 +6,7 @@
 import pytest
 from loguru import logger
 
-from models.demos.mobilenetv2.tests.pcc.test_mobilenetv2 import MOBILENETV2_BATCH_SIZE
+from models.demos.mobilenetv2.common import MOBILENETV2_BATCH_SIZE
 from models.perf.device_perf_utils import check_device_perf, prep_device_perf_report, run_device_perf
